additinal moving around
