+++ src/domain/shared/components/EntityConfirmDeleteDialog.tsx
