+++ src/domain/space/admin/OpportunitySettings/OpportunitySettingsView.tsx
@@ -8,7 +8,7 @@ import { useNotification } from '@/core/ui/notifications/useNotification';
 import { useDeleteSpaceMutation, useSpacePrivilegesQuery } from '@/core/apollo/generated/apollo-hooks';
 import { AuthorizationPrivilege } from '@/core/apollo/generated/graphql-schema';
 import DeleteIcon from '@/domain/journey/space/pages/SpaceSettings/icon/DeleteIcon';
-import EntityConfirmDeleteDialog from '@/domain/journey/space/pages/SpaceSettings/EntityConfirmDeleteDialog';
+import EntityConfirmDeleteDialog from '@/domain/shared/components/EntityConfirmDeleteDialog';
 import { useSubSpace } from '@/domain/journey/subspace/hooks/useSubSpace';
 import PageContent from '@/core/ui/content/PageContent';
 import { useSpace } from '@/domain/space/context/useSpace';
