+++ src/domain/space/icons/SpaceIconByLevel.tsx
